✨ feat: rename home-header and home-footer to app-header and app-footer
