+++ frontend/src/components/app-footer.tsx
